// front-end/src/container/userModules/paymentToDoList/paymentToDoList.types.ts
export interface PaymentData {
  cash_retention_type: string;
  claim_type?: string;
  client_supplier_id: number;
  input_date?: any;
  client_supplier_name?: NamedCurve;
  compulsory_attachment_ids: string[];
  contract_date: string | null;
  list_status: string | null;
  contract_id: number;
  contract_name: string;
  is_retention_confirmed: boolean;
  matched_transactions: string[];
  memo: string;
  optional_attachment_ids: string[];
  amount: number;
  payment_amount: number;
  payment_claim_id: number;
  payment_date: string;
  payment_from_account: string | any | null;
  payment_from_account_name?: string;
  payment_id: number;
  payment_to: string;
  payment_to_account: number;
  payment_type: string;
  project_date: string | null;
  project_id: number;
  project_name: string;
  // retention_account: string | null;
  retention_account_name?: string | null;
  retention_account_number?: number | null;
  retention_amount: number | null;
  retention_release_date: string | null;
  status: string;
  third_party_payment_reason: string | null;
  total_amount: number;
  due_date: any;
  payment_to_account_bsb_number: string;
  payment_to_account_name: string;
  payment_to_account_number: string;
  claim_amount?: number;
  is_late?: boolean;
  sub_payment_type?: string;
}